Add clear button
rect at the bottom of the view that zeroes out the image when
clicked.

Currently the size of the button shrinks as the number of image
rows increases because I went with a very simple implementation.
In the future, I'll make the button a percentage and/or an HTML
element.