Move digit recognition to javascript
I implemented the neural net stuff in Javascript, and set up the Elm
app to send images to JS over a port and get predictions back over
a port.
Performance is much better and it doesn't explode the call stack
and crash in some browsers anymore.
I hope the issues will be resolved in Elm 0.19 so I can do this
in pure Elm.

To run:
    cd <repository root directory>
    elm-make src/App.elm --output=main.js

then open mnist.html in a browser.