Add option to get predictions from remote server
Remote option uses WebSockets. Note that I hard-coded localhost
for the server for now. Also, sending image data happens on
mouse move events, so the expected behavior is to send a ton of
data rapidly. I will probably want to do some rate limiting in
the future.

I kept the option to use the JS implementation with the intention
of eventually using it as a fallback if the websocket connection
fails.