Make InteractiveMnistPlot a bit more configurable
Instead of storing the model and calling a specific recognition function
within the class, the class now stores a function that takes the data as
an argument and is responsible for preprocessing, running input through
the model, and postprocessing the output.

This allows use with different model architectures without modifying the
class.